+++ website/site.js
@@ -3,7 +3,7 @@
     version: '0.1.0',
     designDocs: '22',
     examples: '195',
-    integrationTests: '196',
+    integrationTests: '197',
     differentialTests: '6',
     unitTests: '2',
   });